+++ Changes.md
@@ -1,3 +1,4 @@
+* #740 - Validate rubric ranges when parsing a challenge description
 * #1324 - Add weighted rank to Championship category in finalist scheduling
 * #1315 - Fixed load error with finalist scheduling
 * #1323 - Better handling of display remote control errors